chore: satisfy CI lint and typing gates
Ruff (13 findings + repo-wide format) and mypy --strict (58 errors across
26 files) had never run against this tree: annotate the streaming event
generators, type the JSONB columns and SQL param dicts, stop reusing
loop variables for different row types, and route DELETE intervals
through bound make_interval parameters instead of interpolation. Also
declare ignore_missing_imports for ldap3, defusedxml and fitz. No
behaviour change: 477 tests still green.

+++ backend/app/jobs/registry.py
@@ -35,7 +35,7 @@ class RetryPolicy:
     backoff_seconds: int = 60
 
     def delay_for(self, attempt: int) -> int:
-        return self.backoff_seconds * (2 ** max(0, attempt - 1))
+        return int(self.backoff_seconds * (2 ** max(0, attempt - 1)))
 
 
 _POLICIES: dict[str, RetryPolicy] = {}
